Cleanup: Move IntervalType from merge-tree to sequence
